feat: Add comment functionality to post page
The code changes include adding new files and modifying existing files to implement the comment functionality on the post page. This includes adding a new route for fetching comments, creating components for comment input, comment more button, and delete comment dialog. The changes also include integrating and configuring the uploadthing library.

Note: The recent user commits and recent repository commits are not directly related to the current code changes.

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -55,19 +55,46 @@ export async function POST(
       error:'Unauthorized'
     },{status:401})
 
-    await prisma.like.upsert({
-      where:{
-        userId_postId:{
+    const post = await prisma.post.findUnique({
+      where:{id:postId},
+      select:{
+        userId:true
+      }
+    })
+
+    if(!post) return NextResponse.json({
+      error:'Post not found'
+    },{status:404})
+
+    await prisma.$transaction([
+      prisma.like.upsert({
+        where:{
+          userId_postId:{
+            postId,
+            userId:loggedInUser.id
+          }
+        },
+        create:{
           postId,
           userId:loggedInUser.id
-        }
-      },
-      create:{
-        postId,
-        userId:loggedInUser.id
-      },
-      update:{}
-    })
+        },
+        update:{}
+      }),
+      ...(loggedInUser.id !== post.userId 
+        ? [
+          prisma.notification.create({
+            data:{
+              type:'LIKE',
+              recipientId:post.userId,
+              postId,
+              issuerId:loggedInUser.id
+            }
+          })
+        ] 
+        : [] 
+      )
+    ])
+    
     return NextResponse.json({message:'Success'})
   }
   catch (error) {
@@ -88,12 +115,33 @@ export async function DELETE(
       error:'Unauthorized'
     },{status:401})
 
-    await prisma.like.deleteMany({
-      where:{
-        postId,
-        userId:loggedInUser.id
+    const post = await prisma.post.findUnique({
+      where:{id:postId},
+      select:{
+        userId:true
       }
     })
+
+    if(!post) return NextResponse.json({
+      error:'Post not found'
+    },{status:404})
+
+    await prisma.$transaction([
+      prisma.like.deleteMany({
+        where:{
+          postId,
+          userId:loggedInUser.id
+        }
+      }),
+      prisma.notification.deleteMany({
+        where:{
+          issuerId:loggedInUser.id,
+          recipientId:post.userId,
+          postId,
+          type:'LIKE'
+        }
+      })
+    ])
 
     return new Response()
   }
